+++ devices/mpu6500.h
@@ -61,8 +61,8 @@ enum MPU_DEVTYPE {
 * @brief self-test output data
 */
 struct mpu_stdata {
-	float ax, ay, az;	///< accelerometer x, y and z self-test values
-	float gx, gy, gz;	///< gyroscope x, y and z self-test values
+	double ax, ay, az;	///< accelerometer x, y and z self-test values
+	double gx, gy, gz;	///< gyroscope x, y and z self-test values
 };
 
 /**
